+++ src/transformations/smooth.jl
@@ -12,36 +12,39 @@ end
 _recenter(xn) = (offset = sum(xn) / length(xn); (xn .- offset, offset))
 
 function (l::SmoothAnalysis)(input::ProcessedLayer)
+    tx = position_transform(input.axis_transforms, Lines, input.attributes, 2, 1)
+    ty = position_transform(input.axis_transforms, Lines, input.attributes, 2, 2)
+    scales_active = !isempty(input.axis_transforms)
     if isnothing(l.interval)
         output = map(input) do p, n
             p, _ = _drop_missing_nan_rows(p, n)
             x, y = p
-            xn = to_unitless_numerical(x)
-            yn = to_unitless_numerical(y)
+            xn = to_transformed_numerical(x, tx)
+            yn = to_transformed_numerical(y, ty)
             xn_c, x_offset = _recenter(xn)
             model = Loess.loess(xn_c, yn; l.span, l.degree)
             x̂n = collect(range(extrema(xn)..., length = l.npoints))
             ŷn = Loess.predict(model, x̂n .- x_offset)
-            x̂ = from_unitless_numerical(x̂n, x)
-            ŷ = from_unitless_numerical(ŷn, y)
+            x̂ = from_transformed_numerical(x̂n, x, tx)
+            ŷ = from_transformed_numerical(ŷn, y, ty)
             return (x̂, ŷ), (;)
         end
         plottype = Makie.plottype(output.plottype, Lines)
-        return ProcessedLayer(output; plottype, label = :prediction)
+        return tag_scale_aesthetics(ProcessedLayer(output; plottype, label = :prediction), scales_active)
     else
         output = map(input) do p, n
             p, _ = _drop_missing_nan_rows(p, n)
             x, y = p
-            xn = to_unitless_numerical(x)
-            yn = to_unitless_numerical(y)
+            xn = to_transformed_numerical(x, tx)
+            yn = to_transformed_numerical(y, ty)
             xn_c, x_offset = _recenter(xn)
             model = Loess.loess(xn_c, yn; l.span, l.degree)
             x̂n = collect(range(extrema(xn)..., length = l.npoints))
             pred = Loess.predict(model, x̂n .- x_offset; interval = l.interval, level = l.level)
-            ŷ = from_unitless_numerical(pred.predictions, y)
-            lower = from_unitless_numerical(pred.lower, y)
-            upper = from_unitless_numerical(pred.upper, y)
-            x̂ = from_unitless_numerical(x̂n, x)
+            ŷ = from_transformed_numerical(pred.predictions, y, ty)
+            lower = from_transformed_numerical(pred.lower, y, ty)
+            upper = from_transformed_numerical(pred.upper, y, ty)
+            x̂ = from_transformed_numerical(x̂n, x, tx)
             return (x̂, ŷ, x̂, lower, upper), (;)
         end
 
@@ -59,7 +62,7 @@ function (l::SmoothAnalysis)(input::ProcessedLayer)
             end, plottype = Band, label = :ci, attributes = dictionary([:alpha => 0.15])
         )
 
-        return ProcessedLayers([bandlayer, lineslayer])
+        return tag_scale_aesthetics(ProcessedLayers([bandlayer, lineslayer]), scales_active)
     end
 end